Move source files to macOS and windows directories
